Updated : Playlist functions Added : Description, tags and privacy for playlist Added : Runtime for playlist Added : functions_playlist.php Added : Actions for select queries Added : View playlist page

+++ upload/includes/classes/actions.class.php
@@ -87,6 +87,8 @@ class cbactions
 	 */
 	function init()
 	{
+        global $cb_columns;
+
 		$this->report_opts = array
 		(
 		lang('inapp_content'),
@@ -97,6 +99,19 @@ class cbactions
 		lang('disturbing'),
 		lang('other')		
 		);
+
+        $fields = array( 'playlist_id', 'playlist_name', 'description', 'tags', 'category',
+                         'played', 'privacy', 'total_comments', 'total_items', 'runtime',
+                         'last_update', 'date_added', 'first_item', 'playlist_type', 'cover' );
+
+        $cb_columns->object( 'playlists' )->register_columns( $fields );
+
+        $fields = array(
+            'playlist_item_id', 'object_id', 'playlist_id', 'playlist_item_type', 'userid',
+            'date_added'
+        );
+
+        $cb_columns->object( 'playlist_items' )->register_columns( $fields );
 	}
 
 	/**
@@ -416,7 +431,133 @@ class cbactions
 		else
 			return 0;
 	}
-	
+
+    function load_basic_fields( $array = null ) {
+
+        if ( is_null( $array ) ) {
+            $array = $_POST;
+        }
+
+        $title = $array[ 'playlist_name' ];
+        $description = $array[ 'description' ];
+        $tags = $array[ 'tags' ];
+        $privacy = $array[ 'privacy' ];
+
+        $fields = array(
+            'title' => array(
+                'title' => lang( 'Playlist Name' ),
+                'type' => 'textfield',
+                'name' => 'playlist_name',
+                'id' => 'playlist_name',
+                'db_field' => 'playlist_name',
+                'value' => $title,
+                'required' => 'yes',
+                'invalid_err' => lang( 'Please decide a name for your playlist' )
+            ),
+            'description' => array(
+                'title' => lang ( 'Description' ),
+                'type' => 'textarea',
+                'name' => 'description',
+                'id' => 'description',
+                'db_field' => 'description',
+                'value' => $description
+            ),
+            'tags' => array(
+                'title' => lang( 'Tags' ),
+                'type' => 'textfield',
+                'name' => 'tags',
+                'id' => 'tags',
+                'db_field' => 'tags',
+                'value' => $tags
+            ),
+            'privacy' => array(
+                'title' => 'Privacy',
+                'type' => 'dropdown',
+                'name' => 'privacy',
+                'id' => 'privacy',
+                'db_field' => 'privacy',
+                'value' => array(
+                    'public' => lang( 'Public' ),
+                    'private' => lang( 'Private' ),
+                    #'unlisted' => lang( 'Unlisted' )
+                ),
+                'default_value' => 'public',
+                'checked' => $privacy
+            )
+        );
+
+        return $fields;
+    }
+
+    function load_other_options ( $array = null ) {
+
+        if ( is_null( $array ) ) {
+            $array = $_POST;
+        }
+
+
+        $allow_comments = $array[ 'allow_comments' ];
+        $allow_rating = $array[ 'allow_rating' ];
+
+        $fields = array(
+            'allow_comments' => array(
+                'title' => lang( 'Allow Comments' ),
+                'id' => 'allow_comments',
+                'type' => 'radiobutton',
+                'name' => 'allow_comments',
+                'db_field' => 'allow_comments',
+                'value' => array(
+                    'no' => lang( 'No' ),
+                    'yes' => lang( 'Yes' )
+                ),
+                'default_value' => 'yes',
+                'checked' => $allow_comments
+            ),
+            'allow_rating' => array(
+                'title' => lang( 'Allow Rating' ),
+                'id' => 'allow_rating',
+                'type' => 'radiobutton',
+                'name' => 'allow_rating',
+                'db_field' => 'allow_rating',
+                'value' => array(
+                    'no' => lang( 'No' ),
+                    'yes' => lang( 'Yes' )
+                ),
+                'default_value' => 'yes',
+                'checked' => $allow_rating
+            )
+        );
+
+        return $fields;
+    }
+
+
+    function load_playlist_fields( $array = null ) {
+
+        if ( is_null( $array ) ) {
+            $array = $_POST;
+        }
+
+        $basic = $this->load_basic_fields( $array );
+        $other = $this->load_other_options( $array );
+
+        $fields = array_merge( $basic, $other );
+
+        $group = array(
+            'basic' => array(
+                'group_id' => 'basic_fields',
+                'group_name' => 'Basic Details',
+                'fields' => $basic
+            ),
+            'other' => array(
+                'group_id' => 'other_fields',
+                'group_name' => 'Other Options',
+                'fields' => $other
+            )
+        );
+
+        return $group;
+    }
 	
 	/**
 	 * Function used to create new playlist
@@ -474,11 +615,63 @@ class cbactions
 	/**
 	 * Function used to get playlist
 	 */
-	function get_playlist($id,$user=NULL)
+	function get_playlist( $id, $user = null )
 	{
-		global $db;
-		
-		$user_cond;
+		global $db, $cb_columns;
+
+
+        $fields = array(
+            'playlists' => $cb_columns->object( 'playlists' )->temp_add( 'rated_by,voters,rating,allow_rating,allow_comments' )->get_columns()
+        );
+
+
+        $fields[ 'users' ] = $cb_columns->object( 'users' )->temp_remove('usr_status,user_session_key')->get_columns();
+
+
+
+        $query = "SELECT ".table_fields( $fields )." FROM ".table( 'playlists' );
+        $query .= " LEFT JOIN ".table( 'users' )." ON playlists.userid = users.userid";
+
+        $query .= " WHERE playlists.playlist_id = '$id'";
+
+        if ( !is_null( $user ) and is_numeric( $user ) ) {
+            $query .= " AND playlists.userid = '$user' ";
+        }
+
+        $query .= " LIMIT 1";
+
+        $data = cb_do_action( 'get_playlist', array( 'query' => $query ) );
+
+        if ( $data ) {
+            return $data;
+        }
+
+        $data = select( $query );
+
+        if ( isset( $data[ 0 ] ) and !empty( $data[ 0 ] ) ) {
+            $data = $data[ 0 ];
+
+            if ( !empty( $data[ 'first_item' ] ) ) {
+                $first_item = json_decode( $data[ 'first_item' ], true );
+                if ( $first_item ) {
+                    $data[ 'first_item' ] = $first_item;
+                }
+            }
+
+            if ( !empty( $data[ 'cover' ] ) ) {
+                $cover = json_decode( $data[ 'cover' ], true );
+                if ( $cover ) {
+                    $data[ 'cover' ] = $cover;
+                }
+            }
+
+            return $data;
+        }
+
+
+        return false;
+
+        /*
 		if($user)
 			$user_cond = " AND userid='$user'";
 			
@@ -487,6 +680,7 @@ class cbactions
 			return $result[0];
 		else
 			return false;
+        */
 	}
 	
 	
@@ -495,11 +689,13 @@ class cbactions
 	 */
 	function add_playlist_item($pid,$id)
 	{
-		global $db;
-		
+		global $db, $cb_columns;
+
+        $playlist = $this->get_playlist( $pid );
+
 		if(!$this->exists($id))
 			e(sprintf(lang("obj_not_exists"),$this->name));
-		elseif(!$this->get_playlist($pid))
+		elseif( !$playlist )
 			e(lang("playlist_not_exist"));
 		elseif(!userid())
 			e(lang('you_not_logged_in'));
@@ -507,10 +703,44 @@ class cbactions
 			e(sprintf(lang('this_already_exist_in_pl'),$this->name));
 		else
 		{
+
+            $video = get_video_basic_details( $id, true );
+
+            cb_do_action( 'add_playlist_item', array( 'playlist' => $playlist, 'object' => $video, 'object_type' => $this->type ) );
+
+            if ( !error() ) {
+
+                $fields = array(
+                    'object_id' => $id,
+                    'playlist_id' => $pid,
+                    'date_added' => now(),
+                    'playlist_item_type' => $this->type,
+                    'userid' => userid()
+                );
+
+                /* insert item */
+                $db->insert( tbl( $this->playlist_items_tbl ), array_keys( $fields ), array_values( $fields ) );
+
+                /* update playlist */
+                $fields = array(
+                    'last_update' => now(),
+                    'runtime' => '|f|runtime+'.$video[ 'duration' ],
+                    'first_item' => '|no_mc|'.json_encode( $video ),
+                    'total_items' => '|f|total_items+1'
+                );
+
+                $db->update( tbl( 'playlists' ), array_keys( $fields ), array_values( $fields ), " playlist_id = '".$pid."' " );
+
+                e( sprintf( lang( 'this_thing_added_playlist' ), $this->name ), "m" );
+                return $video;
+            }
+
+            /*
 			$db->insert(tbl($this->playlist_items_tbl),array("object_id","playlist_id","date_added","playlist_item_type","userid"),
 											array($id,$pid,now(),$this->type,userid()));
 			e(sprintf(lang('this_thing_added_playlist'),$this->name),"m");
 			return $db->insert_id();
+            */
 		}
 	}
 	
@@ -521,29 +751,120 @@ class cbactions
 	function delete_playlist_item($id)
 	{
 		global $db;
-		$item = $this->playlist_item($id);		
+
+		$item = $this->playlist_item( $id, true );
+
 		if(!$item)
 			e(lang("playlist_item_not_exist"));
 		elseif($item['userid']!=userid() && !has_access('admin_access'))
 			e(lang("you_dont_hv_permission_del_playlist"));
 		else
 		{
-			$db->delete(tbl($this->playlist_items_tbl),array("playlist_item_id"),array($id));
-			e(lang("playlist_item_delete"),"m");
+            $video = get_video_basic_details( $item[ 'object_id' ] );
+
+            if ( !$video ) {
+                e( lang( "playlist_item_not_exist" ) );
+                return false;
+            }
+
+            cb_do_action( 'delete_playlist_item', array( 'playlist' => $item, 'object' => $video ) );
+
+            /* Remove item */
+			$db->delete( tbl( $this->playlist_items_tbl ),array( "playlist_item_id" ),array( $id ) );
+
+
+            /* Update playlist */
+            $fields = array(
+                'last_update' => NOW(),
+                'runtime' => $item[ 'runtime' ] - $video[ 'duration' ],
+                'total_items' => $item[ 'total_items' ] - 1
+            );
+
+            if ( $fields[ 'runtime' ] <= 0 ) {
+                $fields[ 'runtime' ] = 0;
+            }
+
+            if ( $fields[ 'total_items' ] <= 0 ) {
+                $fields[ 'total_items' ] = 0;
+            }
+
+            if ( $this->is_item_first( $item, $item[ 'object_id' ] ) ) {
+                $fields[ 'first_item' ] = '|no_mc|'.json_encode( array() );
+            }
+
+
+            $db->update( tbl( 'playlists' ), array_keys( $fields ), array_values( $fields ), " playlist_id = '".$item[ 'playlist_id' ]."' " );
+
+			e( lang( "playlist_item_delete" ), "m" );
+
+            return true;
 		}
 	}
-	
+
+    function is_item_first ( $details, $check_id ) {
+
+        if ( !isset( $details[ 'first_item' ] ) ) {
+            return false;
+        }
+
+        $decode = json_decode( $details[ 'first_item' ], true );
+
+        if ( !isset( $decode[ 'videoid' ] ) ) {
+            return false;
+        }
+
+        if ( $decode[ 'videoid' ] == $check_id ) {
+            return true;
+        }
+
+        return false;
+    }
+
 	/**
 	 * Function used to check weather playlist item exists or not
 	 */
-	function playlist_item($id)
+	function playlist_item( $id, $join_playlist = false )
 	{
-		global $db;
+		global $db, $cb_columns;
+
+        $data = cb_do_action( 'get_playlist_item', array( 'playlist_item_id' => $id ) );
+
+        if ( isset( $data ) and $data ) {
+            return $data;
+        }
+
+
+        $fields = array(
+            'playlist_items' => $cb_columns->object( 'playlist_items' )->get_columns()
+        );
+
+        if ( $join_playlist == true ) {
+            $fields[ 'playlists' ] = $cb_columns->object( 'playlists' )->temp_change( 'date_added', 'playlist_added' )->get_columns();
+        }
+
+        $query = "SELECT ".table_fields( $fields )." FROM ".table( 'playlist_items' );
+
+        if ( $join_playlist == true ) {
+            $query .= " LEFT JOIN ".table( 'playlists' )." ON playlist_items.playlist_id = playlists.playlist_id";
+        }
+
+        $query .= " WHERE playlist_items.playlist_item_id = '$id' LIMIT 1";
+
+        $data = select( $query );
+
+        if ( $data ) {
+            return $data[ 0 ];
+        } else {
+            return false;
+        }
+
+        /*
 		$result = $db->select(tbl($this->playlist_items_tbl),"*"," playlist_item_id='$id' ");
 		if($db->num_rows>0)
 			return $result[0];
 		else
 			return false;
+        */
 	}
 	
 	/**
@@ -565,25 +886,79 @@ class cbactions
 	/**
 	 * Function used to update playlist details
 	 */
-	function edit_playlist($params)
+	function edit_playlist( $array = null )
 	{
 		global $db;
-		$name = mysql_clean($params['name']);
-		$pdetails = $this->get_playlist($params['pid']);
+
+        if( is_null( $array ) ) {
+            $array = $_POST;
+
+        }
+
+		$name = mysql_clean($array['name']);
+		$pdetails = $this->get_playlist( $array['pid'] ? $array['pid'] : $array['list_id'] );
 		
 		if(!$pdetails)
 			e(lang("playlist_not_exist"));
 		elseif(!userid())
 			e(lang("you_not_logged_in"));
-		elseif(empty($name))
-			e(lang("please_enter_playlist_name"));
-		elseif($this->playlist_exists($name,userid()) && $pdetails['playlist_name'] !=$name)
-			e(sprintf(lang("play_list_with_this_name_arlready_exists"),$name));
 		else
 		{
-			$db->update(tbl($this->playlist_tbl),array("playlist_name"),
-									  array($name)," playlist_id='".$params['pid']."'");
+
+            $upload_fields = $this->load_playlist_fields( $array );
+            $fields = array();
+
+            foreach( $upload_fields as $group ) {
+
+                $fields = array_merge( $fields, $group[ 'fields' ] );
+
+            }
+
+            validate_cb_form( $fields, $array );
+
+            if( !error() ) {
+
+                foreach($fields as $field)
+                {
+                    $name = formObj::rmBrackets($field['name']);
+                    $val = $array[ $name ];
+
+                    if($field['use_func_val'])
+                        $val = $field['validate_function']($val);
+
+                    if(is_array($val))
+                    {
+                        $new_val = '';
+                        foreach($val as $v)
+                        {
+                            $new_val .= "#".$v."# ";
+                        }
+                        $val = $new_val;
+                    }
+                    if(!$field['clean_func'] || (!function_exists($field['clean_func']) && !is_array($field['clean_func'])))
+                        $val = ($val);
+                    else
+                        $val = apply_func($field['clean_func'],sql_free('|no_mc|'.$val));
+
+                    if(!empty($field['db_field']))
+                        $query_values[ $name ] = $val;
+                }
+
+                if ( has_access( 'admin_access' ) ) {
+
+                    if ( isset( $array[ 'played' ] ) and !empty( $array[ 'played' ] ) ) {
+                        $query_values[ 'played' ] = $array[ 'played' ];
+                    }
+
+                }
+
+                $db->update( tbl( 'playlists' ), array_keys( $query_values ), array_values( $query_values ), " playlist_id = '".$pdetails[ 'playlist_id' ]."' " );
+            }
+
+			/*$db->update(tbl($this->playlist_tbl),array("playlist_name"),
+									  array($name)," playlist_id='".$params['pid']."'");*/
 			e(lang("play_list_updated"),"m");
+
 		}
 	}
 	
@@ -611,23 +986,103 @@ class cbactions
 	/**
 	 * Function used to get playlists
 	 */
-	function get_playlists()
+	function get_playlists( $params = array() )
 	{
-		global $db;
-		$result = $db->select(tbl($this->playlist_tbl),"*"," playlist_type='".$this->type."' AND userid='".userid()."'");
-		
-		if($db->num_rows>0)
-			return $result;
-		else
-			return false;
+        global $cb_columns;
+
+        $fields = array(
+            'playlists' => $cb_columns->object( 'playlists' )->get_columns()
+        );
+
+        $order = $params[ 'order' ];
+        $limit = $params[ 'limit' ];
+
+        $main_query = $query = "SELECT ".table_fields( $fields )." FROM ".table( 'playlists' );
+        $condition = "playlists.playlist_type = 'v'";
+
+        if ( !has_access( 'admin_access' ) ) {
+            $condition .= ( $condition ) ? " AND " : "";
+            $condition .= "playlists.privacy = 'public'";
+        } else {
+            if ( $params[ 'privacy' ] ) {
+                $condition .= ( $condition ) ? " AND " : "";
+                $condition .= " playlists.privacy = '".mysql_clean( $params[ 'privacy' ] )."' ";
+            }
+        }
+
+        if ( $params[ 'category' ] ) {
+            $condition .= ( $condition ) ? " AND " : "";
+            $condition .= " playlists.category = '".$params[ 'category' ]."' ";
+        }
+
+        if ( $params[ 'include' ] ) {
+            $ids = is_array( $params[ 'include' ] ) ? $params[ 'include' ] : explode( ',', $params[ 'include' ] );
+
+            if ( is_array( $ids ) and !empty( $ids ) ) {
+                $condition .= ( $condition ) ? " AND " : "";
+                $ids = implode( ",", array_map( 'trim', $ids ) );
+                $condition .= " playlists.playlist_id IN ($ids) ";
+            }
+        }
+
+        if ( $params[ 'exclude' ] ) {
+            $ids = is_array( $params[ 'exclude' ] ) ? $params[ 'exclude' ] : explode( ',', $params[ 'exclude' ] );
+
+            if ( is_array( $ids) and !empty( $ids ) ) {
+                $condition .= ( $condition ) ? " AND " : "";
+                $ids = implode( ",", array_map( 'trim', $ids ) );
+                $condition .= " playlists.playlist_id NOT IN ($ids) ";
+            }
+        }
+
+        if ( $params[ 'date_span' ] ) {
+            $condition .= ( $condition ) ? " AND " : "";
+            $column = ( $params[ 'date_span_column' ] ) ? trim( $params[ 'date_span_column' ] ) : 'playlists.date_added';
+
+            $condition .= cbsearch::date_margin( $column, $params['date_span'] );
+        }
+
+        if ( $params[ 'last_update' ] ) {
+            $condition .= ( $condition ) ? " AND " : "";
+            $condition .= cbsearch::date_margin( 'playlists.last_update', $params['last_update'] );
+        }
+
+        if( $params[ 'user' ] ) {
+            $condition .= ( $condition ) ? " AND " : "";
+            $condition .= " playlists.userid = '".$params[ 'user' ]."' ";
+        }
+
+        if ( $condition ) {
+            $query .= " WHERE ".$condition;
+        }
+
+        $order = " ORDER BY ".( $order ? trim( $order ) : "playlists.date_added DESC");
+        $limit = ( $limit ) ? " LIMIT $limit ": "";
+
+
+        $query .= $order.$limit;
+        $results = select( $query );
+
+        if ( !empty( $results ) ) {
+            return $results;
+        }
+
+        return false;
 	}
 	
 	/**
 	 * Function used to get playlist items
 	 */
-	function get_playlist_items($pid)
+	function get_playlist_items( $playlist_id, $order = null, $limit = -1 )
 	{
-		global $db;
+		global $db, $cb_columns;
+
+        $fields = array(
+            'playlist_items' => $cb_columns->object( 'playlist_items' )->get_columns(),
+            'playlists' => $cb_columns->object( 'playlists' )->temp_remove( 'first_item' )->get_columns(),
+            'video' => $cb_columns->object( 'videos' )->get_columns()
+        );
+
 		$result = $db->select(tbl($this->playlist_items_tbl),"*","playlist_id='$pid'");
 		if($db->num_rows>0)
 			return $result;
